mae_Retrain/requirements.txt
pytorch # the later version is also OK
torchvision==0.8.2
timm==0.4.12
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
# deepspeed==0.4.0
scipy
